requirements.txt
chromadb
sentence_transformers
python-dotenv
wikipedia
torch
transformers
huggingface_hub
